.github/scripts/requirements.txt
# Dependências para o AIRT - Relatório Automático de Testes
# Scripts de extração, análise e publicação de resultados de testes

# Biblioteca para integração com Confluence
atlassian-python-api>=3.41.0

# Bibliotecas para parsing e manipulação de dados
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Biblioteca para requisições HTTP
requests>=2.31.0

# Bibliotecas para logging e manipulação de arquivos
pathlib2>=2.3.7; python_version < "3.4"

# Bibliotecas opcionais para melhorias
# colorama>=0.4.6  # Para logs coloridos (opcional)
# rich>=13.0.0     # Para interface rica (opcional) 